adjust time to secs (int32) and nanos (uint32). Index as combined int64

+++ seerep-msgs/core/timestamp.h
@@ -7,8 +7,8 @@ namespace seerep_core_msgs
 {
 struct Timestamp
 {
-  int64_t seconds;
-  int64_t nanos;
+  int32_t seconds;
+  uint32_t nanos;
 };
 } /* namespace seerep_core_msgs */
